+++ tests/admin_interface.rs
@@ -44,7 +44,6 @@ async fn admin_interface() {
         "--authentication-request-timeout=5s",
         "--http-request-timeout=5s",
     ]);
-    dbg!(&config);
     tokio::spawn(async move { entrypoint(config).await });
     if timeout(Duration::from_secs(5), async {
         while let Err(_) = TcpStream::connect("127.0.0.1:18022").await {
@@ -148,6 +147,7 @@ async fn admin_interface() {
     });
     if timeout(Duration::from_secs(3), async move {
         // 4a. Validate header, system information, and HTTP tab data
+        dbg!("4a");
         let search_strings: Vec<Regex> = [
             r"Sandhole admin v\d+\.\d+\.\d+",
             r"System information",
@@ -170,6 +170,7 @@ async fn admin_interface() {
             }
         }
         // 4b. Switch tabs and validate SSH tab data
+        dbg!("4b");
         writer
             .write(&b"\t"[..])
             .await
@@ -196,6 +197,7 @@ async fn admin_interface() {
             }
         }
         // 4c. Switch tabs again and validate TCP tab data
+        dbg!("4c");
         writer
             .write(&b"\t"[..])
             .await
@@ -222,6 +224,7 @@ async fn admin_interface() {
             }
         }
         // 4d. Go back one tab
+        dbg!("4d");
         writer
             .write(&b"\x1b[Z"[..])
             .await
@@ -248,6 +251,7 @@ async fn admin_interface() {
             }
         }
         // 4e. Quit the admin interface with Ctrl-C (ETX)
+        dbg!("4e");
         writer
             .write(&b"\x03"[..])
             .await